backend/requirements.txt
fastapi
uvicorn
sqlmodel
transformers
torch
pillow
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
pytest
pytest-cov
httpx
argon2-cffi
loguru
slowapi
python-dotenv
